+++ kommander/src/main/java/com/wokdsem/kommander/Cancelable.java
@@ -1,0 +1,7 @@
+package com.wokdsem.kommander;
+
+interface Cancelable {
+	
+	void cancel();
+	
+}