+++ kommander/src/main/java/com/wokdsem/kommander/RunnableAction.java
@@ -1,14 +1,7 @@
 package com.wokdsem.kommander;
 
-class RunnableAction<T> implements Runnable {
-
-	private enum RunnableState {
-		NEW,
-		RUNNING,
-		COMPLETED,
-		CANCELED
-	}
-
+class RunnableAction<T> implements Runnable, Cancelable {
+	
 	private Action<T> action;
 	private Deliverer deliverer;
 	private Response.OnCompleted<T> onCompleted;
@@ -17,15 +10,15 @@ class RunnableAction<T> implements Runnable {
 	private Thread executor;
 	private Throwable error;
 	private T result;
-
-	RunnableAction(RunnableActionBundle<T> bundle) {
+	
+	RunnableAction(RunnableActionBundle<T> bundle, Deliverer deliverer) {
 		this.action = bundle.action;
-		this.deliverer = bundle.deliverer;
+		this.deliverer = deliverer;
 		this.onCompleted = bundle.onCompleted;
 		this.onError = bundle.onError;
 		this.state = RunnableState.NEW;
 	}
-
+	
 	@Override
 	public final void run() {
 		Action<T> runnableAction;
@@ -50,7 +43,7 @@ class RunnableAction<T> implements Runnable {
 			}
 		}
 	}
-
+	
 	private synchronized void onResult(T result) {
 		if (state != RunnableState.CANCELED) {
 			if (onCompleted != null) {
@@ -67,14 +60,14 @@ class RunnableAction<T> implements Runnable {
 			}
 		}
 	}
-
+	
 	private synchronized void deliverResult() {
 		if (state != RunnableState.CANCELED) {
 			onCompleted.onCompleted(result);
 			release(true);
 		}
 	}
-
+	
 	private synchronized void onError(Throwable error) {
 		if (state != RunnableState.CANCELED) {
 			if (onError != null) {
@@ -91,15 +84,16 @@ class RunnableAction<T> implements Runnable {
 			}
 		}
 	}
-
+	
 	private synchronized void deliverError() {
 		if (state != RunnableState.CANCELED) {
 			onError.onError(error);
 			release(true);
 		}
 	}
-
-	synchronized void cancel() {
+	
+	@Override
+	public synchronized void cancel() {
 		if (state == RunnableState.NEW || state == RunnableState.RUNNING) {
 			if (executor != null) {
 				executor.interrupt();
@@ -109,7 +103,7 @@ class RunnableAction<T> implements Runnable {
 			release(false);
 		}
 	}
-
+	
 	private void release(boolean isFullyCompleted) {
 		this.state = isFullyCompleted ? RunnableState.COMPLETED : RunnableState.CANCELED;
 		onCompleted = null;
@@ -117,5 +111,12 @@ class RunnableAction<T> implements Runnable {
 		result = null;
 		error = null;
 	}
-
+	
+	private enum RunnableState {
+		NEW,
+		RUNNING,
+		COMPLETED,
+		CANCELED
+	}
+	
 }